pyproject.toml
[project]
name = "codeit-part3-team4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "camelot-py>=1.0.9",
    "docling>=2.72.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.13.2",
    "img2table>=1.4.2",
    "ipykernel>=7.1.0",
    "marker-pdf>=1.6.1",
    "matplotlib>=3.10.8",
    "numpy<2.0",
    "olefile>=0.47",
    "openai>=2.17.0",
    "openpyxl>=3.1.5",
    "paddleocr[doc-parser]>=3.4.0",
    "paddlex>=3.4.2",
    "pandas<3.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.2",
    "tokenizers>=0.22.2",
    "transformers>=4.57.6",
]
